pyproject.toml
[tool.poetry]
name = "cra-sion"
version = "0.0.1"
description = "A client-server based game inspired by Super Crate Box."
authors = ["Frexom <imfrexom@gmail.com>", "YamaXuna <>"]
license = "GNU AGPLv3"
readme = "README.md"
packages = [{include = "cra_sion"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
paho-mqtt = "^1.6.1"
urllib3 = "^2.1.0"
pillow = "^10.1.0"
idna = "^3.4"
charset-normalizer = "^3.3.2"
certifi = "^2023.7.22"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["."]
